Remove unused CI workflows and outdated configuration files
Deleted obsolete GitLab and GitHub CI/CD workflows, IntelliJ code style settings, and other project-specific configuration files that are no longer required. Updated references in the OpenAPI documentation to reflect branding changes from "HERE" to "XYZmaps".
