--- .github/CODEOWNERS
@@ -1,1 +1,0 @@
-* @heremaps/xyz-hub
