--- .github/workflows/test.yaml
@@ -1,24 +1,0 @@
-name: Run XYZ Hub tests
-
-on: [push, pull_request, workflow_dispatch]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Cache local Maven repository
-        uses: actions/cache@v2
-        with:
-          path: ~/.m2/repository
-          key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml') }}
-          restore-keys: |
-            ${{ runner.os }}-maven-
-      - name: Setup Java
-        uses: actions/setup-java@v1
-        with:
-          java-version: '8'
-      - name: Start the XYZ Hub stack
-        run: mvn clean install -Pdocker -DskipTests=true -DdockerComposeFile=docker-compose-dynamodb.yml
-      - name: Run tests
-        run: mvn verify -DskipTests=false
