+++ xyz-hub-service/src/main/java/com/here/xyz/hub/rest/FeatureApi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2020 HERE Europe B.V.
+ * Copyright (C) 2017-2023 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,14 +28,12 @@ import static com.here.xyz.hub.rest.ApiParam.Query.SKIP_CACHE;
 import static io.vertx.core.http.HttpHeaders.ACCEPT;
 
 import com.here.xyz.events.ContextAwareEvent.SpaceContext;
-import com.here.xyz.events.DeleteFeaturesByTagEvent;
 import com.here.xyz.events.GetFeaturesByIdEvent;
 import com.here.xyz.events.ModifyFeaturesEvent;
 import com.here.xyz.events.TagsQuery;
 import com.here.xyz.hub.rest.ApiParam.Path;
 import com.here.xyz.hub.rest.ApiParam.Query;
 import com.here.xyz.hub.task.FeatureTask.ConditionalOperation;
-import com.here.xyz.hub.task.FeatureTask.DeleteOperation;
 import com.here.xyz.hub.task.FeatureTask.IdsQuery;
 import com.here.xyz.hub.task.ModifyFeatureOp;
 import com.here.xyz.hub.task.ModifyFeatureOp.FeatureEntry;
@@ -179,24 +177,13 @@ public class FeatureApi extends SpaceBasedApi {
     final SpaceContext spaceContext = getSpaceContext(context);
 
     //Delete features by IDs
-    if (featureIds != null && !featureIds.isEmpty()) {
+    if (!featureIds.isEmpty()) {
       Map<String, Object> featureModification = Collections.singletonMap("featureIds", new ArrayList<>(featureIds));
 
       executeConditionalOperationChain(false, context, responseType, IfExists.DELETE, IfNotExists.RETAIN, true,
           ConflictResolution.ERROR, Collections.singletonList(featureModification), spaceContext);
     }
-
-    //Delete features by tags
-    else if (!tags.isEmpty()) {
-      if (checkModificationOnSuper(context, spaceContext))
-        return;
-      DeleteFeaturesByTagEvent event = new DeleteFeaturesByTagEvent();
-      if (!tags.containsWildcard()) {
-        event.setTags(tags);
-      }
-      new DeleteOperation(event, context, responseType)
-          .execute(this::sendResponse, this::sendErrorResponse);
-    } else {
+    else {
       context.fail(
           new HttpException(HttpResponseStatus.BAD_REQUEST, "At least one identifier should be provided as a query parameter."));
     }
