--- xyz-models/src/main/java/com/here/xyz/events/DeleteFeaturesByTagEvent.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2017-2022 HERE Europe B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-package com.here.xyz.events;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonTypeName(value = "DeleteFeaturesByTagEvent")
-@Deprecated
-public final class DeleteFeaturesByTagEvent extends QueryEvent<DeleteFeaturesByTagEvent> {
-
-}
