+++ xyz-models/src/main/java/com/here/xyz/events/SelectiveEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2022 HERE Europe B.V.
+ * Copyright (C) 2017-2023 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,6 @@ public class SelectiveEvent<T extends SelectiveEvent> extends ContextAwareEvent<
   private boolean force2D;
   private String ref;
   private long minVersion;
-
   private String author;
 
   @SuppressWarnings("unused")
