+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/DatabaseHandler.java
@@ -35,7 +35,6 @@ import com.here.xyz.XyzSerializable;
 import com.here.xyz.connectors.ErrorResponseException;
 import com.here.xyz.connectors.SimulatedContext;
 import com.here.xyz.connectors.StorageConnector;
-import com.here.xyz.events.DeleteFeaturesByTagEvent;
 import com.here.xyz.events.Event;
 import com.here.xyz.events.HealthCheckEvent;
 import com.here.xyz.events.IterateFeaturesEvent;
@@ -490,6 +489,11 @@ public abstract class DatabaseHandler extends StorageConnector {
             else if(event.getOperation() == Operation.UPDATE)
                 //Update HistoryTrigger to apply maxVersionCount.
                 updateHistoryTrigger(event, maxVersionCount, compactHistory, isEnableGlobalVersioning);
+        }else if(event.getSpaceDefinition() != null && !event.getSpaceDefinition().isEnableHistory()){
+            if (event.getOperation() == Operation.CREATE) {
+                //Create Space Table
+                ensureSpace();
+            }
         }
 
         new ModifySpace(event, this).write();
@@ -796,21 +800,6 @@ public abstract class DatabaseHandler extends StorageConnector {
       ids.addAll(deletes.keySet());
 
       return ids;
-    }
-
-    @Deprecated
-    protected XyzResponse executeDeleteFeaturesByTag(DeleteFeaturesByTagEvent event) throws SQLException {
-        boolean includeOldStates = event.getParams() != null
-                && event.getParams().get(INCLUDE_OLD_STATES) == Boolean.TRUE;
-
-        final SQLQuery searchQuery = GetFeaturesByBBox.generateSearchQueryBWC(event);
-        final SQLQuery query = SQLQueryBuilder.buildDeleteFeaturesByTagQuery(includeOldStates, searchQuery);
-
-        //TODO: check in detail what we want to return
-        if (searchQuery != null && includeOldStates)
-            return executeQueryWithRetry(query, this::oldStatesResultSetHandler,false);
-
-        return new FeatureCollection().withCount((long) executeUpdateWithRetry(query));
     }
 
     private boolean canRetryAttempt() throws Exception {
