+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/PSQLXyzConnector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2022 HERE Europe B.V.
+ * Copyright (C) 2017-2023 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,6 @@ import static com.here.xyz.responses.XyzError.ILLEGAL_ARGUMENT;
 
 import com.here.xyz.connectors.ErrorResponseException;
 import com.here.xyz.events.DeleteChangesetsEvent;
-import com.here.xyz.events.DeleteFeaturesByTagEvent;
 import com.here.xyz.events.GetFeaturesByBBoxEvent;
 import com.here.xyz.events.GetFeaturesByGeometryEvent;
 import com.here.xyz.events.GetFeaturesByIdEvent;
@@ -226,23 +225,6 @@ public class PSQLXyzConnector extends DatabaseHandler {
     }
     finally {
       logger.info("{} Finished " + event.getClass().getSimpleName(), traceItem);
-    }
-  }
-
-  @Override
-  @Deprecated
-  protected XyzResponse processDeleteFeaturesByTagEvent(DeleteFeaturesByTagEvent event) throws Exception {
-    try{
-      logger.info("{} Received DeleteFeaturesByTagEvent", traceItem);
-      if (config.getDatabaseSettings().isReadOnly()) {
-        return new ErrorResponse().withStreamId(streamId).withError(XyzError.NOT_IMPLEMENTED)
-                .withErrorMessage("ModifyFeaturesEvent is not supported by this storage connector.");
-      }
-      return executeDeleteFeaturesByTag(event);
-    }catch (SQLException e){
-      return checkSQLException(e, config.readTableFromEvent(event));
-    }finally {
-      logger.info("{} Finished DeleteFeaturesByTagEvent", traceItem);
     }
   }
