+++ xyz-psql-connector/src/test/java/com/here/xyz/psql/PSQLWriteIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2022 HERE Europe B.V.
+ * Copyright (C) 2017-2023 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -186,10 +186,6 @@ public class PSQLWriteIT extends PSQLAbstractIT {
 
         assertUpdate(updateRequest, updateResponse, false);
         LOGGER.info("Update feature tested successfully");
-
-        // =========== DELETE FEATURES ==========
-        invokeLambdaFromFile("/events/DeleteFeaturesByTagEvent.json");
-        LOGGER.info("Delete feature tested successfully");
     }
 
     @Test
